Use geocoder gem for doing the distance based database searches
Still using geokit gem for doing the actual geocoding as that should
hopefully still work under Rails 3. However, the geocoder gem currently
has a bug where the calculated distance is always returned in miles even
when the search units are set to kilometres. For the time being doing
the conversions back to km inline.