Don't test rendering the promo because it's random
I toyed with the idea of testing if the rendered email contained
either promo but that would mean you've got a random element in your
tests, i.e. it may fail or succeed depending on which random partial
gets chosen.

I'd still like a way to test the render_random_promo method.