Replace rotating promo with a far simpler one for all our projects
I started this by wanting to cut down the size of the promo as it's
too loud right now. Maybe OK for our new project but for everyday
alerts it's not nice for the user to have this Big Honking Promo on
every email.

Anyway, I started cutting down the size and realised I could simplify
this right down to a promo for all the projects in one sentence. I
liked it and also like the fact it cuts down the complexity of the
implementation. Let's see how it performs and go from there.