Completely remove mobile version of the site
About one in every three visits to the site is on a mobile device.
Four out of five of those are on an iPad or iPhone. Those devices, and
the vast majority of the other devices used, have a very capable
browser. We're currently showing them a significantly chopped down
version of the site, including removing the key commenting feature.

This makes me sad.

Note also that we're already planning to move to a responsive design
(see ticket).