Fix regression where partial can't be found, fixes #778
When the applications/show is called from the comment controlled,
it searches for partials in the views/comments directory.

Using an absolute path (from the view directory) fixes this as
rails can always find the partial no matter which controller the
applications/show view is called from.

This also adds an integration test for the situation that
this bug was discovered in.