Move privacy notice to where it’s more relevant
If the commenter has options as to where their comment will be sent
it doesn't make sense to inform them that the comment will be posted
on this page at the top. It’s more relevant at the point that they’re
going to post their comment.

This only effects the version of the form with councillors.

If they can only write to the council, then it still makes sense to
have this notice when the user is told that their comment will go
to the council, because they could assume it just goes to council.