Use a more intuitive verb and make comment button personal
The term 'create' here refers to the record being created in the
database. This is not relevant to people making a comment.

The verb 'post' would be clearer here to describe the comment being
posted to the page and also to the council. Post is the term
used by Facebook to describe making a status update as well, so
people may be comfortable with this concept.

Also refer to 'your comment' rather than just 'comment' to
re-enforce the person’s sense of ownership of their action.