Indent councillor list so it appears as a sub option of its heading
As pointed out by @henare, having the list of councillors appear
inline with (or as a continuation of) the 'who should this go to?'
makes these two separate tasks appear as one. This is confusing.

This commit simply indents the councillor list, so it appears
as a child of the 'One of your elected local councillors' option,
which is how it should be understood.