Make label simpler and more encouraging
We want people to feel good and be empowered to comment on local
development. We should be friendly and encouraging to them as they
take action.

This change makes the process sound less bureaucratic and a bit more
exciting. You're 'having your say' rather than 'explaining what you
think'

In our commenters survey (see #770), two of the three top response
to "Why did you comment on a development applications" were
"to raise my concerns about a development application"
and "to suggest changes to a development application before it is approved".
Respondents could select multiple answers and 62% selected the
former, while 42% selected the latter. 51% selected
"to suggest that my council reject a development application"
and 10% selected to "suggest that my council approve a
development application".

So lots of people are commenting for reasons other than to suggest
an application should be approved or not.

This change makes the comment label less proscriptive as to what
should go in the comment. I think this better reflects range of
motivations people described in the survey.