Make placeholder text more helpful by focusing on the 'why'
Why should people be polite, concise and to the point?
In their own interest, it is because it will make their comment
more effective with the council. A polite and concise comment is
more readable and understandable that a shouty one that waffles on.
This realistically means that council staff are more likley to
understand, respect and share its content.

I think it's a bit more fun, and less 'stuffy' to simply say:
'Effective comments are ...' than 'Please be ...'.

It also reads like a tip rather than a request.