Replace 'concise' with 'clear'
'Concise' is quite close in meaning to 'to the point' but is a bit
more formal and less friendly. 'Concise' isn't adding much here.

Removing removing 'concise' and just leaving:
'Effective comments are polite and to the point.'
Sounds a bit abrupt and unfriendly.

Replacing 'concise' with 'clear' adds more meaning, and is a simpler
word that will help people write an effective comment.