Fix bug where comment was wrapped in <li> but wasn't in a list
This fixes #801 "Comment report form has a stray bullet point"
which was caused by wrapping the comment in an <li> when it
wasn't part of a list.

This adds an option to the comment template to say whether or not
it is part of a list. This means we can continue to use the handy
rails helper `render @comment`.

I tried another solution first, by directly rendering the
`comment_item` template on the report page, but decided this was
a bad solution. If we can't use the `render @comment` syntax to
simply render a comment, because it depends on an additional html
pattern, then we're undermining a really handy rails convention.

I think this solution is better because it gives us a very clear
way to describe how we want our comment to be rendered.

Fixes #801