Don't show expired warning before form, if you can write to councillors
From discussion on GitHub:
https://github.com/openaustralia/planningalerts/pull/788#issuecomment-143655893

> If the commenting period is over I'm warned in bold above the form,
> "...not be officially considered...". For some reason this felt out of
> place. Now I think I know what it is - the form intro talks about
> official submissions but then the form allows me to do something else.
> I think we need to fix this clashing.

It’s confusing to get warned about being too late before you've
even decided whether to send an official submission.

This moves that warning onto the option to make an official submission
by writing to the council.

So you've presented with the information at the point that it's actually
useful to you.