Customise the expired message when you can also write to councillors
The default 'official comment period expired' message is too off-putting
in the situation when you have other options, like writing to your
councillors.

This forks the text from the on_notice_text helper and front loads the
important info telling you exactly what you're 'too late for':
> to have your comment officially considered when they decide on this application

This makes the commenter’s situation and options clearer.

I'm not sure if this change should be made in the helper because it may
be more important to discourage people from commenting if the period is
expired and they don't have the option to write to councillors.