Focus on the benefit to the user when directing them to write clear comments
From the discussion at
https://github.com/openaustralia/planningalerts/pull/797#issuecomment-147614539

In the commenter's survey responses there's a clear pattern that people
are concerned about whether their comment makes an impact:

> I have no idea if the comments will be listened to or what impact they
> will have if any

> I do get the 'Form Letter' from Council but I'm not in any way
> convinced they listen.

> I have no real way of knowing whether my concerns are given any
> attention by local council.

People want their comments to “be listened to”. We think this is more
likely to happen if their comments are polite, clear and succinct than
if they're rude, confused and rabbling.

If we focus the comment text around this benefit, and use the language
that people making comments use, then hopefully they'll listen to us ;)
This makes the form more useful to them as it's helping them achieve the
impact they're looking for.