Reduce prominence of donations declaration warning
It's important the people know they made need to declare their political
donations in their comments, but this information only relates to a very small set
of people who have donated. It also reminds people about the boring
bureaucracy of council submission and I think will be off putting in
this current primary position in the form.

This keeps this information visible but moves it into helper text
beneath the input so it’s not in the path leading to writing your comment,
but something you check during/after writing it.