Let user know their comment is public during the action they're focused on (Form with councillors only)
Rather than adding more elements to focus on, we should simplify the form
by giving you important privacy information when it's relevant to you,
rather than putting it in a separate paragraph you're likely to skip
over. Rather than just posting a 'comment' you should be clicking a button
to post a 'public comment'.

I've only made this change in for the form with councillors because
this was how the form worked when it was a separate form.
This iteration is meant to be a simple move from a separate form for
writing to councillors, to picking who you'll send to as part of
a making a regular comment.

On the separate form version the button to submit read:

> 'Post your public comment ...'

and there was no:

> Your name and comment will be posted publicly here.

So removing this text fixes a regression that slipped in
when merging these forms together.