Halve the hight of the comment text input textarea
The size of an input field should default to the intended length of the
text to be inputed. By my quick look, most comments are either roughly
half the length of the text fields default size, or are much longer. The
textarea can be stretched to be taller by the user.

By making it bigger we're making the form longer, and seem more involved
than it needs to be. We're also suggesting a length that most people
don't seem to think is necessary. This is in conflict with our direction
that people be 'to the point' in writing their comments.

Making it shorter will make the form simpler and appear a bit more
approachable and help people scan the inputs—making the form a little
more usable.

Fixes #815