Re-enforce that comments are sent to authority to impact decision
Add text to show where the comment was sent to:

1. Make it clearer what happens with these comments, as distinct
from messages to councillors. Add the date stamp so people know
when they got it. This makes the comment seem more official.
2. Re-enforce that these comments to get sent to the authority,
that they will impact the decision.

Also add that the comment was 'posted' here, to also make the concept
that comments are on PlanningAlerts for PlanningAlerts users, not just
the council, as well.

Important TODO on this:
> Change this to show datestamp in useful timezone to the user, not UTC.
> Maybe local time of application or local time of user?

I'm not sure the best way to do this yet, but it's not super important
for this prototype.