Compile applications js into site wide file, save a request (refactor only)
The applications js file, that adds special behaviours to application
show pages, is only <2kb compressed but we're making an additional http
request to fetch it. I assume this was done to save other users from
this weight. But, this template is the most visited, so most users are
hitting this page anyway and having to download this as a separate file.
So the cost of an additional request is greater than this tiny weight.

This commit sets rails to compile applications.js into the main global
js file.

This is a recommendation from ySlow
https://developer.yahoo.com/performance/rules.html#num_http=

Also, see this handy StackOverflow answer:
http://stackoverflow.com/questions/9855545/http-requests-vs-file-size