Move commenter name above comment in email for consistency with site
On the website, the names of commenters are positioned above what they
say. The email version was not update to match. They should be
consistent so people learn the pattern. Also, we moved the name up to
emphasise the humans taking action, this change should be made in emails
too.

This is also a usability improvement because currently you have to
scroll to the end of the comment to know which application it refers
too. Currently the comment has no context before you start reading.

Fixes #817