Increase the contrast of input text
The contrast of input text is reduced, making it slightly softer
and less clear.

There’s no obvious reason for this, and it incrementally makes using
the forms for of a task for some users.

I noticed this in usability testing on where a person was squinting
as they checked the comment they had inputed.