Always refer to councillors with their title
Most people wont know who their councillors are by name.

In user testing we saw that not recognising the name in the comment
footer as a councillor, lead the viewer to think that this was the
name of the person who had sent the message.

More discussion on the testing here:
https://github.com/openaustralia/planningalerts/issues/799#issuecomment-152380457