Make wording more explicit
In user testing, the view was confused by this text, they weren't
sure if the person named was sending or receiving the message.

Using 'delivered' rather than 'sent' makes the intended meaning
more explicit. It’s also less passive, we might have “sent” it, but
who knows what happened in the delivery. If we say “delivered”, we’re
making it clearer that the person has this message.