Replace wrapping links on comments with link on time and address
The wrapping link renders inconsistently in in different browsers because
of the links, such as the report link, inside the wrapping link.
Interactive content inside interactive content is not allowed in
the HTML spec.

This replaces it with a link to the application if the address of
the comment is visible, and links to the actual comment where the
time is shown.