Highlight difference between types of comments by emphasising the people
This increases the weight of the names of the people involved in
a comment. For comments to planning authorities, this just means
the comment author. For messages to councillors, this means the
author and the local councillor. Users should notice the two people
vs one person pattern. The aim is to re-enforced one of the key
conceptual differences between the types of comments:
Messages to councillors are a dialog, where as comments to th
authority are one way.
This 2 vs 1 variation should be noticeable at a glance.

Earlier we’d looked at making this change to emphasise the people
involved in commenting, but decided not to do it because we were
concerned that it would diminish the focus on the comment text in
relation to the author name too much. After some user testing,
I've noticed that people currently barely focus on this comment
header section, and are very interested in the comment text. I
don’t think we risk pushing the balance too far towards the names.

Read more about this previous discussion here:
https://github.com/openaustralia/planningalerts/pull/792#issuecomment-143959085

Relates to:
https://github.com/openaustralia/planningalerts/issues/799