Use generic label for option because people wont recognise authority by name
People shouldn’t have to know the name of the planning authority to
decide whether or not they want to send a comment to them.

Throughout PlanningAlerts we use the generic term 'planning authority'
to refer to these institutions.

Instead of using the authority name for the option label,
name the authority in the label description text so that
people are still informed of precisely who their message will go to.

This should be done in the description in addition to repeating the
term 'planning authority' to make sure people make that connection.

Fixes #829