Give some notice that you can choose who to write to above first input
Currently, unless you see the options about who you'll send you comment
to, you’re likely to make at least small changes to your comment if you
decide to write to a councillor. This is because the options of who to
send it to come as a surprise.

We originally moved these options down in the form to let people get
straight into commenting, because in user testing the person got stuck
deciding who to write to before they had thought about what they wanted
to say. I think that’s still a good choice, but we could could some kind
of simple heads up/signpost, at the top of the form. I think this would
be particularly helpful for new commenters.