Remove duplication between heading and label
The comment form intro is overly complex with duplication between
the section heading 'Add your own comment' and
'Have you say on this application'. Also the whitespace between the
heading, introduction paragraphs, and label doesn’t lead into the form
because there’s a bigger gap between the intro par and the form is
bigger than the space between the heading and other page sections.

This commit removes the current heading 'Add your own comment' and
makes the first form label the heading, with the intro par underneath.

This simplifies the top of the form.