Move notice into button text
The notice that your comment and name will be publicly posted on
PlanningAlerts in buried in the introduction par. In the form version with
councillors included we removed this notice and made the submit button
read 'Post your public comment' so that the information is in a more
obvious place, and is connected to the action in relates to, posting the
comment.

This repeats that change for the current versions of the comment form so
that both versions are the same.

This does remove the information that specifically your name and comment
will be public. But I think that if this is a “public comment” and we
say we don’t show your address or email in the form, then it’s clear
enough what will be public.