Distinguish commenting options more in the introduction text
People need to understand that they have different options for who
their comment should be sent to, and how they are different.

This changes the introduction text to use different verbs for the
different types of responses you can make.

“Post your comment for the planning authority” is meant to re-enforce
the concept that this is a public comment that is 'for' the planning
authority to use—it doesn’t just go in the bin, it matters.
I like 'for' over 'to' here because it hints that
the comments are actually useful and welcomed by the authority, it’s
less passive.

“—write to one of your elected local councillors” is meant to re-enforce
that this is a less formal message, in the familiar concept of
‘write to your MP’.