Show a link to the application with the comment on the report page
This enables people to go back to the application and also gives
more context to the comment as they are writing their report.
They may want to reference the address in their report so this is
handy.

This changes the comment partial to enable showing a link to the
application when the comment isn't part of a list and updates tests
accordingly.