Simplify comment footer text
Having the time here is confusing because we already have the time
the comment was posted, in a different format, on the comment. It
leads you to think 'Is this posted date different from when the
person commented?' It was introduces another verb, 'posted'.

We’re removed this second date stamp in this commit.

We also removed the reference to the authority by their name,
e.g. 'City of Sydney', because the user doesn't need to know who
these people are and wont even necessarily recognise them as the
planning authority.

For the new situation where there are comments that are sent to
different places, this differentiates this kind of comment.

We’re also deploying this for authorities that wont have councillor
comments yet to re-enforce where this is sent and that it
will make an impact. We can do this more simply with less text and avoid
confusion.