Clarity 'official commenting period' notice
This makes the notice focused on the important impact to the person,
will their comment be official considered or not?

It also makes it clearer that this is the planning authority’s time
limit, and that you can comment outside of it (or write to your
councillors) but knowing how that impacts you.

This makes it clear that there is only a time limit on comments
if you want to write to the planning authority.