Simplify sentence and don’t reply on 'letter' analogy
It needs to be clear that these messages to councillors will be public.
Currently we're leaning on the letter analogy, but when you send a
letter there is a presumption that the letter stays closed until it gets
to its intended target. We don’t want people to think that about these
comments so we should try and avoid that analogy.