Re-enforce that your comment and name will be public
We know some people miss this text currently and post a comment
without realising that their name and comment will be public.
We’ve added 'public' to the button, and remove this text, but now
after talking with @henare , we’re adding it back to really hit this
home.

It also adds a nice friendly, personal touch to the intro text.