Revert "Give comments in emails colours like in the website for clarity and consistency"
The colour styles are too different from the style of the emails
currently, which is quite distinct from the style of the website.

For the MVP at least, we wont make the comments in the email look too
much like the website, because they’ll look very out of place in the
email.

This reverts commit efd4cdcf25ef3c565c88b1149fee89c2c946bf4f.