Rename method
It was actually activeadmin that made me think about this. It
automatically looks for a `display_name` method and uses that in the
interface.

This made me realise that calling this method `display_name` was a bit
inaccurate - it's really a special prefixed name, you don't want to
use it as a display name everywhere.