Make replies section header text consistent with other sections
We had original not included the alert area 'within x distance of x address',
because it felt repetitive and unnecessary. But, when we look at this in
context, in the email alert, the variation from the other headers is
confusing. “Why is it talking about comments near an address but replies
near me?”

While I think we should extract this address information out into some
kind of header for the address, I think the email is more familiar and
understandable if all the headers are consistent in how they refer to
the location.

If we made them all “near you” then there would be no mention of the
address that the alert was for in the email, so I think adding the
address back to the replies section is the better small step.