Make the user pick who to send a comment to if options are available
Currently, if the comment form contains the options for who it should be
sent to, and the user fills out all the form inputs except who it should
go to, then the comment is accepted and sent to the planning authority.

This is definitely a bug, because the user has not indicated that this
is what they want.

This adds a custom validation to ensure a receiver is selected if the
options are available.