Persist the state of the councillors list toggler
This means that if a user opens the councillor list before submitting
the form, and are then returned to the page because of a validation error,
the list will still be open as they would expect.

This is must important if they have selected a councillor and hit an
error in another part of the form. With persisting the state of the
toggler, the list would be closed and they wouldn't be able to see that
they have a councillor selected.