Rework comment validations to enable existing comments to be updated
If the way this was initially implemented in 9cadb3d , was requiring a
special attribute to be set for us to update a comment. That would mean
adding new code to the admin views just to complete existing tasks.

I've reworked the validation on Comment that ensures the user selects
who their comment should go to if they have the options, so that it
doesn't impact updating existing comments.

This also simplifies the model and takes advantage of helpers from
Active Record Validations.