Extract Class for comment form
This extracts the special logic for the comment form out from the
comment model, into a new model, controller etc.

The logic for handling and validating the comment form with councillors
was getting very complex when mixed up in all the regular comment
business logic. This makes both classes clearer and isolates all the
special validation.