Don't take an address for comments to councillor
Providing your address wont necessarily help you get want you need when
you write to a councillor.

A planning authority is accountable in ways that a councillor, an
individual is not. Individual councillors should not automatically be given
the street address of people writing to them.

To simplify view code and to reduce the risk of a data leak, this
implementation removes an address if the user has provided one.

There's some duplication in this implementation between the validation
on the comment form object CreateComment and on the comment model. It
feels like we need the check in both places to me, but if a way can be
found to remove the code repetition but keep the behaviour, the could be
a helpful simplification.

Closes #843
https://github.com/openaustralia/planningalerts/issues/843