Rename class for better routes
You can't change the path name for a create action route, so I'm changed
the name of the class from CreateComment to AddComment. Now you get the
route: /applications/1353/add_comments .

I think this “add” is much nicer for users seeing the path than
“create”, which sounds very databasy. This could still be better but I
can't think of the words right now.

I wanted to change this back to the old 'comments' to match what it was
like before the comment form object class was extracted, but as far as I
can tell that's not possible. Still, I think '/applications/1353/add_comments'
is better than add '/applications/1353/comments'.