Fix link to comments admin view from report admin view
This fixes a long running bug caused by an old problem in ActiveAdmin.
https://github.com/activeadmin/activeadmin/issues/301

ActiveAdmin took the /admin/comments/ path for a comment feature we
aren't using. This meant we had to use /admin/application_comments/
unlike the reset of our admin routes, which match the public model
routes after /admin/.

This meant that in the reports view, the links to the comment being
reported assumed it was looking for /admin/comments/:id , which would
404.

The fix, changing the ActiveAdmin comments registration name:

    config.comments_registration_name = "AdminComment"

–is described here:
https://github.com/activeadmin/activeadmin/issues/301#issuecomment-37889465

This commit renames the url helpers throughout the site so they all keep
working as expected.

Since these links only effect admins I don't think redirects are
necessary.