Remove commenter’s email from report emails to avoid potential data leak
Often administrators (including me) just reply to these emails to the
reporter explaining their decision. At the moment they are forced to
remember to remove the email of the commenter to avoid leaking their
private email address to the reporter.

This commit removes the email and a test to make sure it is not
accidentally added back in.

Fixes #857
https://github.com/openaustralia/planningalerts/issues/857