Use built-in browser validation to provide immediate feedback on form input
If the browser supports it, provide immediate feedback on form inputs
without reloading the page. This is a win because page reloads take
time, and jump the user to a new part of the page (or a new page in many
of our cases), making it difficult to reorientate and find the form
errors.

This just switches on the inbuilt browser validation, we could do a more
custom version using javascript as a progressive enhancement on to of
this.

See browser support of the required attribute:
http://caniuse.com/#feat=form-validation