Make charts represent data in more detail
The current curvy interpolated line of the area chart isn't appropriate
for our data which is has lots of sudden drop offs (when a scraper fails
and it stops collecting applications) and sudden peaks (such as when
there are hundreds of comments on a popular application). The
interpolation smooths out these lines, and compresses the troughs and
peaks, making it hard to see when comments or applications suddenly drop
off. This compression might be useful if we had less frequent data
sampling and wanted to interpolate the in-between values, but we don't
have that need. This is made worse with the detailed y-axis ticks that give the
impression that the data is also granular.

This commit changes to more of a bar style chart, that is highly
accurate but still useful for seeing broad trends.