Fix performance tests time zone problem
The implementation of the performance page data queries uses just UTC
for consistency. The tests are mixed up, part UTC and part local.

This is causing inregular test results and failures:
https://travis-ci.org/openaustralia/planningalerts/jobs/100675316

Make all the performance tests use UTC and set the current time more
accurately in Timecop for a bit more clarity of what's going on.