Clip the bold line at the 0 values so 0 values are more promient
One problem with the charts originally was that you couldn't see 0
values vary well. These are important because they could reveal a
problem with a scraper or comments dropping off after a change.

This makes the bold edge of the chart values disappear if the value is
below 0, which draws attention to these sections.

This implementation is based on:
http://bl.ocks.org/mbostock/4062844