Show the value and dates in focus to the side of the chart
Currently it's very difficult to tell precisely which week you're focused
on, because the scale can be so wide and not granular. When you're
looking at the chart, trying to understand the impact of a design or
code change, you need to be able to focus in on the dates you're
interested in.

This commit reduces the width of the chart and adds text providing
precise detail of the area you're focused on with the mouse or touch.
If you have no area in focus it defaults to the latest week, as this is
most likely the information you're after when you hit this page.

This design pattern is based on the GDS performance charts at:
https://www.gov.uk/performance/carers-allowance