Make the chart with callout label work with small screens in a basic way
This scales down the chart for smaller screens so they
can be read without having to scroll the screen across a wide area.

To improve the experience for small screens further, the next step with
this is to use js to redraw the chart to be optimised for the current
viewport size. Currently the text scales down in the chart so the axis
become hard to read. I've compensated for this by increasing the type
size a bit, but this can't be pushed far because really the number of
ticks in each axis needs to adapt.

Still, this is a step forward for users on small screens and makes the
charts much more accessible to them.