Add test to check that reply text doesn't include any sneaky javascript
Since we're rendering this stuff to html so that users get nice
formatting, I thought it was a good idea to make sure we sanitize
reply.text of any javascript that might slip in.

Currently administrators are entering the replies manually, but that
will hopefully not be the case soon.