Make utility 'hide content' style selector much more specific (fixes #899)
This fixes the bug where the marker for the location of the application
in our embedded Google Map is missing:
https://github.com/openaustralia/planningalerts/issues/899

`aria-hidden` is used in the Google Map to hide the marker for some
reason. So when I added this utility class in
ab680c6dbd526afc193e6b58be475475052c5728 it hid the marker.

This commit reduces the specificity of the hiding style to just target
the use case at hand: hiding the comment address field when a
councillor has been selected.

I was being a bit to clever by setting that to be a global utility
class. Should have done the dumb thing then!