Allow admin to toggle availability of writing to councillors for authority
See issue:
https://github.com/openaustralia/planningalerts/issues/914

In #898 we decided to not let people write to councillors unless we have
full data for all the councillors on the authority.

We want to control whether the commenting option is shown using a manual
control on the admin page for an authority. This is the simple way to
control this as we're still learning about how this feature is used.

Closes #914