Fix test which fails when run in certain order
This test currently fails when run with seed 13341 because it looks for
the 'last email', but that is sometimes the confirm comment email.
We don't care about the order of these emails, so fix this by opening
the email we want by it's subject.

This also adds a TODO to extract the dodgy assignment of 'comment' into
a method where the user creates a comment and confirms it via their
email. This is the better way to resolve this issue.