Show when the comment was confirmed rather than updated
Fixes #930
https://github.com/openaustralia/planningalerts/issues/930

Currently the time we say the person made a comment is the `updated_at`
time. For most comments this ends up being the time the `confirmed`
attribute was updated, i.e. when they confirmed the comment. But when an
Admin updates the comment for some reason, the `updated_at` time changes
and we then wrongly say that is when the person commented.
This has lead to situations where the website says someone commented
days after they actually commented.

This uses a new attribute 'confirmed_at' which is set when the person
confirms their comment.