Add migration to update all existing confirmed comments confirmed_at
This will make the `confirmed_at` value for all existing confirmed
comments the current `updated_at` value. We currently use `updated_at`
as a stand in for the time a comment was confirmed, because most
comments are never updated otherwise. This has caused a bug, where
the time we say someone posted a comment is wrong (#930).
https://github.com/openaustralia/planningalerts/issues/930

When future comments are confirmed, the `confirmed_at` value is set.
See 95c7aa1 and a0169ac for more about how that works.
We need to update previous comments to standardise the existing comments
to the new schema.

The value is updated with a tested Class method:
`.fill_confirmed_at_for_existing_confirmed_comments`