Included comments in alert based on when the were confirmed
Currently comments are included in an alert if they were **updated**
within the cutoff time. This means that if an Admin updates a comment
for some reason, that comment could get sent out again to local
subscribers for a second time. This is really bad for users because
the comment might have been updated for privacy reason, and the
double send draws attention to it.

Using the `confirmed_at` time rather than the `updated_at` time means
that Admins can safely update comments and they wont be mistakenly
included in alerts.