Put the person’s message straight up front in the text email
Currently this email is structured to have a very robotic block with
information about the message in front of the message.

The important content in the email is the message, so let's put that
straight up front and then provide the context afterwards.
This puts the focus on what the user was trying to say.

Relates to #953
https://github.com/openaustralia/planningalerts/issues/953

We're currently only sending these text emails to councillors as WriteIt
wont accept HTML email.