Only the link to view the comment should be bold
We've just made URLs in comments into clickable links in the alert
email. Currently they're styled bold, like the link to view the comment
on PlanningAlerts (the “class to action” of the comment item).

This looks really messy, and adds emphasis to urls people add for no
good reason. Make the URLs not-bold.