Restructure the HTML email to councillors
This moves the person's message to the top of the email
and makes the explanation information more like an email footer.

This makes the email structure much simpler and emphasises the exchange
between the citizen and the councillor.

It also makes the template text a little less personal by removing
the name of the councillor (“Hi Louise Councillor”).
The tone of the email is now more determined by the citizen.