Add the councillors name back into the email
Having the name confirms that it's for them and further personalises
it.

It also provides some help if someone other than the intended person is
looking at it. After looking at more of the email addresses, I'd say
it's common for the incomings get forwarded around to a few people,
particularly the major@ addresses.

I've added this back into the test as having it there picked up it's
removal once already ;-)