On the authority’s applications page, use a local address as example
This helps make this placeholder meaningful to the user and hopefully
will help them better understand the form.

I've used the last application, not the first, so that they don't get
confused and think they need to put in the address of an application
they want to get alerts for.