Show the sign up form on the applications index without an authority
This simplifies the logic around when this form is shown and it would be
helpful to users on the application index anyway.

It also changes the logic around how the placeholder address is chosen
so it will get the last one on this page, rather than the last one
generally. This will be more dynamic which I think will be nice.