Emphasise that the image can be selected + which councillor is selected
This adds a border around the councillor image when you hover, and a
stronger one when the councillor is selected.

This makes it much clearer that you don't have to interact with the tiny
radio button, but can click the whole block. It's also much clearer
which councillor you've selected.

I've used the blue colours to contrast the yellow and match the
link/hover style throughout the site.